requirements.txt
# FastAPI and server (replacing Flask)
fastapi==0.115.0
uvicorn==0.31.0

# Database
sqlalchemy==2.0.35
aiomysql==0.2.0
asyncmy==0.2.9

# Authentication
fastapi-users[sqlalchemy]==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# OAuth
authlib==1.3.2
httpx==0.27.2

# Validation and config
pydantic==2.0.3
python-dotenv==1.0.1

# Templates
jinja2==3.1.4

# Utilities (from existing project)
requests==2.31.0
Pillow==10.1.0
pyttsx3==2.90
python-dateutil==2.8.2
